requirements.txt
fastapi
uvicorn
langchain
langgraph
faiss-cpu
rank_bm25
sentence-transformers
requests
python-dotenv